Выделен user-service, все тесты проходят

+++ core/user-service/src/main/java/ru/practicum/service/UserServiceImpl.java
@@ -1,5 +1,8 @@
-package ru.practicum.service.impl;
+package ru.practicum.service;
 
+import exception.ConflictException;
+import exception.NotFoundException;
+import exception.ValidationException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -9,12 +12,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import ru.practicum.dal.UserRepository;
-import ru.practicum.dto.user.NewUserDto;
-import ru.practicum.dto.user.UserDto;
-import ru.practicum.exceptions.NotFoundException;
-import ru.practicum.mappers.UserMapper;
+import dto.user.NewUserDto;
+import dto.user.UserDto;
+import ru.practicum.mapper.UserMapper;
 import ru.practicum.model.User;
-import ru.practicum.service.UserService;
 
 import java.util.Collection;
 import java.util.List;
@@ -30,29 +31,32 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @Override
     public UserDto newUser(NewUserDto dto) {
+        if (userRepository.findUserByEmail(dto.getEmail()).isPresent())
+            throw new ConflictException("User with email %s exists!".formatted(dto.getEmail()));
         User user = userRepository.save(userMapper.getUser(dto));
         log.info("Added new user with id {}", user.getId());
-        return userMapper.getUserDto(user);
+        return UserMapper.INSTANCE.getUserDto(user);
     }
 
     @Override
     public Collection<UserDto> getAllUsers(List<Long> ids, int from, int size) {
         Pageable pageable = PageRequest.of(from > 0 ? from / size : 0, size, Sort.by(Sort.Direction.ASC, "id"));
         return userRepository.findAllByFilter(ids, pageable).stream()
-                .map(userMapper::getUserDto)
+                .map(UserMapper.INSTANCE::getUserDto)
                 .toList();
     }
 
     @Transactional
     @Override
     public void deleteUser(long userId) {
-        checkUser(userId);
+        if (getUserById(userId) == null)
+            throw new NotFoundException("User with id %d not found!".formatted(userId));
         userRepository.deleteById(userId);
         log.info("User {} deleted", userId);
     }
 
-    private void checkUser(long userId) {
-        userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundException("User with id %d not found!".formatted(userId)));
+    @Override
+    public UserDto getUserById(Long userId) {
+        return UserMapper.INSTANCE.getUserDto(userRepository.findById(userId).orElse(null));
     }
 }
